+++ eras/dijkstra/src/Cardano/Ledger/Dijkstra/Genesis.hs
@@ -1,5 +1,6 @@
 {-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE DerivingVia #-}
+{-# LANGUAGE RecordWildCards #-}
 {-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeFamilies #-}
 {-# OPTIONS_GHC -Wno-orphans #-}
@@ -18,19 +19,25 @@ import Cardano.Ledger.Binary (
 import Cardano.Ledger.Binary.Coders
 import Cardano.Ledger.Core
 import Cardano.Ledger.Dijkstra.Era (DijkstraEra)
+import Cardano.Ledger.Dijkstra.PParams (UpgradeDijkstraPParams)
 import Cardano.Ledger.Genesis (EraGenesis (..))
-import Data.Aeson (FromJSON (..), ToJSON, withObject)
+import Data.Aeson (FromJSON (..), ToJSON, Value (..), withObject)
+import Data.Functor.Identity (Identity)
 import GHC.Generics
 import NoThunks.Class (NoThunks)
 
 -- TODO: Currently it is just a placeholder for all the new protocol parameters that will be added
 -- in the Dijkstra era
 data DijkstraGenesis = DijkstraGenesis
+  { dgUpgradePParams :: !(UpgradeDijkstraPParams Identity DijkstraEra)
+  }
   deriving (Eq, Show, Generic)
   deriving (ToJSON) via KeyValuePairs DijkstraGenesis
 
 instance FromJSON DijkstraGenesis where
-  parseJSON = withObject "DijkstraGenesis" $ \_ -> pure DijkstraGenesis
+  parseJSON = withObject "DijkstraGenesis" $ \obj -> do
+    dgUpgradePParams <- parseJSON (Object obj)
+    pure DijkstraGenesis {..}
 
 instance NoThunks DijkstraGenesis
 
@@ -39,18 +46,23 @@ instance EraGenesis DijkstraEra where
 
 -- TODO: Implement this and use for ToJSON instance
 instance ToKeyValuePairs DijkstraGenesis where
-  toKeyValuePairs DijkstraGenesis = []
+  toKeyValuePairs dg@(DijkstraGenesis _) =
+    let DijkstraGenesis {..} = dg
+     in toKeyValuePairs dgUpgradePParams
 
 instance FromCBOR DijkstraGenesis where
   fromCBOR =
     eraDecoder @DijkstraEra $
       decode $
         RecD DijkstraGenesis
+          <! From
 
 instance ToCBOR DijkstraGenesis where
-  toCBOR DijkstraGenesis =
-    toEraCBOR @DijkstraEra . encode $
-      Rec DijkstraGenesis
+  toCBOR dg@(DijkstraGenesis _) =
+    let DijkstraGenesis {..} = dg
+     in toEraCBOR @DijkstraEra . encode $
+          Rec DijkstraGenesis
+            !> To dgUpgradePParams
 
 instance DecCBOR DijkstraGenesis