+++ eras/dijkstra/testlib/Test/Cardano/Ledger/Dijkstra/ImpTest.hs
@@ -1,12 +1,21 @@
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE MultiParamTypeClasses #-}
+{-# LANGUAGE NumericUnderscores #-}
+{-# LANGUAGE RecordWildCards #-}
 {-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeFamilies #-}
 {-# OPTIONS_GHC -Wno-orphans #-}
 
-module Test.Cardano.Ledger.Dijkstra.ImpTest () where
+module Test.Cardano.Ledger.Dijkstra.ImpTest (
+  exampleDijkstraGenesis,
+) where
 
-import Cardano.Ledger.BaseTypes (EpochInterval (..), addEpochInterval)
+import Cardano.Ledger.BaseTypes (
+  BoundedRational (..),
+  EpochInterval (..),
+  addEpochInterval,
+  knownNonZeroBounded,
+ )
 import Cardano.Ledger.Conway.Governance (ConwayEraGov (..), committeeMembersL)
 import Cardano.Ledger.Conway.Rules (
   ConwayCertPredFailure (..),
@@ -17,16 +26,18 @@ import Cardano.Ledger.Conway.Rules (
 import Cardano.Ledger.Dijkstra (DijkstraEra)
 import Cardano.Ledger.Dijkstra.Core
 import Cardano.Ledger.Dijkstra.Genesis (DijkstraGenesis (..))
+import Cardano.Ledger.Dijkstra.PParams (UpgradeDijkstraPParams (..))
 import Cardano.Ledger.Plutus (SLanguage (..))
 import Cardano.Ledger.Shelley.LedgerState (epochStateGovStateL, nesEsL)
 import Cardano.Ledger.Shelley.Rules (ShelleyDelegPredFailure)
 import qualified Cardano.Ledger.Shelley.Rules as Shelley
+import Data.Maybe (fromJust)
 import Lens.Micro ((%~), (&))
 import Test.Cardano.Ledger.Conway.ImpTest
 import Test.Cardano.Ledger.Dijkstra.Era ()
 
 instance ShelleyEraImp DijkstraEra where
-  initGenesis = pure DijkstraGenesis
+  initGenesis = pure exampleDijkstraGenesis
 
   initNewEpochState = defaultInitNewEpochState $ \nes ->
     nes
@@ -68,3 +79,15 @@ instance InjectRuleFailure "DELEG" ShelleyDelegPredFailure DijkstraEra where
   injectFailure (Shelley.StakeKeyNotRegisteredDELEG c) = StakeKeyNotRegisteredDELEG c
   injectFailure (Shelley.StakeKeyNonZeroAccountBalanceDELEG c) = StakeKeyHasNonZeroRewardAccountBalanceDELEG c
   injectFailure _ = error "Cannot inject ShelleyDelegPredFailure into DijkstraEra"
+
+exampleDijkstraGenesis :: DijkstraGenesis
+exampleDijkstraGenesis =
+  DijkstraGenesis
+    { dgUpgradePParams =
+        UpgradeDijkstraPParams
+          { udppMaxRefScriptSizePerBlock = 1024 * 1024 -- 1MiB
+          , udppMaxRefScriptSizePerTx = 200 * 1024 -- 200KiB
+          , udppRefScriptCostStride = knownNonZeroBounded @25_600 -- 25 KiB
+          , udppRefScriptCostMultiplier = fromJust $ boundRational 1.2
+          }
+    }