+++ eras/dijkstra/testlib/Test/Cardano/Ledger/Dijkstra/TreeDiff.hs
@@ -5,12 +5,19 @@ module Test.Cardano.Ledger.Dijkstra.TreeDiff (
   module Test.Cardano.Ledger.Conway.TreeDiff,
 ) where
 
+import Cardano.Ledger.BaseTypes (StrictMaybe)
 import Cardano.Ledger.Dijkstra (DijkstraEra)
 import Cardano.Ledger.Dijkstra.Core (EraTx (..), EraTxBody (..), PlutusScript)
+import Cardano.Ledger.Dijkstra.PParams (DijkstraPParams)
 import Cardano.Ledger.Dijkstra.TxBody (DijkstraTxBodyRaw)
+import Data.Functor.Identity (Identity)
 import Test.Cardano.Ledger.Conway.TreeDiff (ToExpr)
 
 instance ToExpr (PlutusScript DijkstraEra)
+
+instance ToExpr (DijkstraPParams Identity DijkstraEra)
+
+instance ToExpr (DijkstraPParams StrictMaybe DijkstraEra)
 
 instance ToExpr DijkstraTxBodyRaw